change a few "int msg_len" to "size_t msg_len"; adjust docs
* lib/cpg.c (cpg_mcast_joined):
* lib/evs.c (evs_mcast_joined, evs_mcast_groups):
* man/cpg_initialize.3:
* man/evs_initialize.3:

git-svn-id: http://svn.fedorahosted.org/svn/corosync/trunk@2060 fd59a12c-fef9-0310-b244-a6a79926bd2f

+++ lib/cpg.c
@@ -501,7 +501,7 @@ cs_error_t cpg_mcast_joined (
 	struct iovec iov[64];
 	struct req_lib_cpg_mcast req_lib_cpg_mcast;
 	struct res_lib_cpg_mcast res_lib_cpg_mcast;
-	int msg_len = 0;
+	size_t msg_len = 0;
 
 	error = saHandleInstanceGet (&cpg_handle_t_db, handle, (void *)&cpg_inst);
 	if (error != CS_OK) {
